refactor(entity)!: move and rename core entities
