+++ src/domain/entities/core/persistable.entity.ts
