+++ src/domain/entities/core/persistence-type.entity.ts
