+++ src/domain/ports/storage-provider.port.ts
@@ -1,4 +1,4 @@
-import { PersistenceType } from '@src/domain/entities/persistence-type.entity';
+import { PersistenceType } from '@src/domain/entities/core/persistence-type.entity';
 import { KeyValueStoragePort } from '@src/domain/ports/key-value-storage.port';
 
 export interface StorageProviderPort {
